Move dispatch source vocabulary out of hooks
