+++ src/mindroom/turn_controller.py
@@ -44,13 +44,12 @@ from mindroom.constants import (
     RuntimePaths,
 )
 from mindroom.delivery_gateway import SendTextRequest
+from mindroom.dispatch_source import is_automation_source_kind, is_voice_event
 from mindroom.error_handling import get_user_friendly_error_message
 from mindroom.handled_turns import HandledTurnState
 from mindroom.hooks import (
     build_hook_matrix_admin,
     hook_ingress_policy,
-    is_automation_source_kind,
-    is_voice_event,
     should_handle_interactive_text_response,
 )
 from mindroom.inbound_turn_normalizer import (
